fix: give Undo snackbars a 10 s window; prove undo in flow 02 (BUG-11)
Running the Maestro suite on the Pixel 8 Pro showed flow 03 hitting the
"Add another?" prompt after an Undo — the undone event was still there.
Undo snackbars used SnackbarDuration.Short (4 s); a tap landing as the
snackbar auto-dismisses is reported as Dismissed, which clears the
pending undo without deleting. Too short for a real person too.

All three undo snackbars (Home log, Calendar and History delete) now use
Long (10 s), per Material guidance for actionable snackbars. Flow 02 now
opens History after Undo and asserts no event rows — "snackbar gone" was
also satisfied by a timeout, so it had never proved undo worked.

+++ composeApp/src/commonMain/kotlin/com/ostomate/app/ui/calendar/CalendarScreen.kt
@@ -107,7 +107,8 @@ fun CalendarScreen(viewModel: CalendarViewModel = koinViewModel()) {
             snackbarHostState.showSnackbar(
                 message = eventDeletedMsg,
                 actionLabel = undoLabel,
-                duration = SnackbarDuration.Short,
+                // BUG-11: give the Undo a real 10 s window
+                duration = SnackbarDuration.Long,
             )
         when (result) {
             SnackbarResult.ActionPerformed -> viewModel.undoDelete()
